backend/requirements.txt
google-genai
asyncio
websockets
pydantic
python-dotenv
pytest
